+++ public/apps/client-cli/src/modules/run/run.ts
@@ -3,7 +3,8 @@ import {defineCommand} from 'citty';
 import {useContainer} from '../../services/container';
 import {useConfig} from '../../services/config';
 import {type ClientConfig} from '../../services/config.d';
-import type {Handlers, ParsedName, Categories} from './run.d'
+import type {Handlers} from './run.d'
+import {getContainerName, resolveRawName} from './utils/naming';
 
 const {runContainer} = useContainer();
 const {getConfig} = useConfig();
@@ -21,6 +22,11 @@ export const run = defineCommand({
     description: 'Run tasks, jobs and apps',
   },
   args: {
+    detach: {
+      description: 'Run in background in detached mode',
+      alias: 'd',
+      type: 'boolean',
+    },
     name: {
       description: 'Name of tasks, jobs and apps',
       type: 'positional',
@@ -37,25 +43,35 @@ export const run = defineCommand({
     if (!category || !names.length) {
       return consola.error(`No task, job or app found with name '${rawName}'`);
     }
-    await Promise.all(names.map((name) => handlers[category](name, config)))
+
+    const runOptions: any = {
+      detach: !!args.detach,
+      silentLog: !!args.detach,
+      onSpawn: args.detach ? () => {
+        consola.success('Process Started')
+      }: undefined
+    }
+
+    await Promise.all(names.map((name) => {
+      handlers[category](name, config, {...runOptions, containerName: getContainerName(category, name) })
+    }))
 
   },
 });
 
-async function handleJob(name: string, config: ClientConfig) {
+async function handleJob(name: string, config: ClientConfig, runOptions: any) {
   const handlers = config.jobs[name].tasks.map((val: string | string[]) =>
     () => {
       if(!Array.isArray(val)) {
-        return handleTask(val, config)
+        return handleTask(val, config, runOptions)
       }
       return Promise.all(val.map((taskName: string) => handleTask(taskName, config)))
     }, [])
   handlers.forEach((handler) => handler())
 }
 
-async function handleApp(name: string, config: ClientConfig) {
+async function handleApp(name: string, config: ClientConfig, runOptions: any) {
   const {expose, task} = config.apps[name];
-  const containerName = `bit-app-${name}`
   const ports: string[]  = []
   if(expose) {
     expose.map((exp) => {
@@ -64,10 +80,11 @@ async function handleApp(name: string, config: ClientConfig) {
       }
     })
   }
-  handleTask(task, config, {containerName, ports})
+
+  handleTask(task, config, {...runOptions, ports})
 }
 
-async function handleTask(taskName: string, config: ClientConfig, runOptions?: any) {
+async function handleTask(taskName: string, config: ClientConfig, runOptions: any) {
   const task = config?.tasks[taskName];
   if (!task) {
     throw 'No task found for name: ' + taskName
@@ -76,30 +93,7 @@ async function handleTask(taskName: string, config: ClientConfig, runOptions?: a
   if (!image) {
     throw 'No image found for name: ' + taskName
   }
-  const containerName = `bit-${taskName}-${Date.now()}`
-  runContainer({remove: true, taskName, containerName, image, script: task.script, volumes: ['./:/app'], storeLogs: true, ...runOptions});
-}
-
-
-// ================================> Helpers
-const resolveRawName = (rawName: string, config: ClientConfig): ParsedName => {
-  const categories: Categories[]  = ['jobs', 'apps', 'tasks']
-  const  category =
-    // Prefix Match
-    categories.find((category) =>
-      rawName.startsWith(`${category}:`) && config[category][rawName.replace(`${category}:`, '')]) ||
-    //Item Match
-    categories.find((category) => config?.[category]?.[rawName])
-  const wildCardCategory = categories.find((cat) => cat === rawName)
-
-  let names: string[] = []
-  if (category) {
-    names.push(rawName.replace(category+':', ''))
-  } else if (wildCardCategory) {
-    names = Object.keys(config[wildCardCategory])
-  }
-
-  return  { names, category: category || wildCardCategory }
+  runContainer({remove: true, taskName, image,env: task.env, script: task.script, volumes: ['./:/app'], storeLogs: true, ...runOptions});
 }